% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/import.R
\name{importehars}
\alias{importehars}
\title{An Import Function}
\usage{
importehars(eharsdir)
}
\arguments{
\item{eharsdir}{The directory in which the eHARS person view file can be found}
}
\description{
This function allows you to import data from eHARS into R
}
\keyword{ehars}
\keyword{import}

